+++ pulse.loop/Models/OpticalFrontendReading.swift
@@ -11,8 +11,8 @@ struct OpticalFrontendReading {
     let value: Double
     let date: Date
     
-    init(value: Double, date: Date = Date.now) {
-        self.value = value
+    init(value: Float32, date: Date = Date.now) {
+        self.value = Double(value)
         self.date = date
     }
 }
